LabWork4/zookeeper-two-phase-commit/requirements.txt
kazoo==2.8.0
python-dotenv==0.19.2
